abstract out file validation
